refactor components -> individual directories
