+++ src/components/db/SampleQueries/SampleQueries.js
@@ -1,16 +1,16 @@
 import React, { Component } from 'react';
 import moment from 'moment';
 
-import queryLibrary from '../../api/data/queries/query-library';
-import HalinQuery from '../../api/data/queries/HalinQuery';
-import sentry from '../../api/sentry';
+import queryLibrary from '../../../api/data/queries/query-library';
+import HalinQuery from '../../../api/data/queries/HalinQuery';
+import sentry from '../../../api/sentry';
 
-import DBStats from '../dbstats/DBStats';
-import Spinner from '../ui/scaffold/Spinner/Spinner';
-import Explainer from '../ui/scaffold/Explainer/Explainer';
-import QueryExecutionPlan from './QueryExecutionPlan';
-import QueryStatTable from './queries/QueryStatTable';
-import CSVDownload from '../data/download/CSVDownload';
+import DBStats from '../../dbstats/DBStats';
+import Spinner from '../../ui/scaffold/Spinner/Spinner';
+import Explainer from '../../ui/scaffold/Explainer/Explainer';
+import QueryExecutionPlan from '../QueryExecutionPlan/QueryExecutionPlan';
+import QueryStatTable from '../queries/QueryStatTable';
+import CSVDownload from '../../data/download/CSVDownload';
 
 import {
     Button,
