+++ src/components/diagnostic/DiagnosticPane/DiagnosticPane.css
