+++ src/components/diagnostic/StorageCapacity/StorageCapacity.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
-import CypherDataTable from '../data/CypherDataTable/CypherDataTable';
-import Explainer from '../ui/scaffold/Explainer/Explainer';
-import queryLibrary from '../../api/data/queries/query-library';
+import CypherDataTable from '../../data/CypherDataTable/CypherDataTable';
+import Explainer from '../../ui/scaffold/Explainer/Explainer';
+import queryLibrary from '../../../api/data/queries/query-library';
 import { Message } from 'semantic-ui-react';
-import hoc from '../higherOrderComponents';
+import hoc from '../../higherOrderComponents';
 
 class StorageCapacity extends Component {
     state = {
