+++ src/components/performance/OSPane/OSPane.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import "semantic-ui-css/semantic.min.css";
 import { Grid } from 'semantic-ui-react';
-import PageCache from '../../diagnostic/PageCache';
-import StorageCapacity from '../../diagnostic/StorageCapacity';
+import PageCache from '../../diagnostic/PageCache/PageCache';
+import StorageCapacity from '../../diagnostic/StorageCapacity/StorageCapacity';
 import OSStats from '../OSStats/OSStats';
 import uuid from 'uuid';
