+++ src/components/ui/scaffold/MainLeftNav/MainLeftNav.js
@@ -4,7 +4,7 @@ import { Sidebar, Segment, Menu, Icon, Image, Popup } from 'semantic-ui-react';
 import ClusterOverviewPane from '../../../overview/ClusterOverviewPane/ClusterOverviewPane';
 import PermissionsPane from '../../../configuration/PermissionsPane/PermissionsPane';
 import SettingsPane from '../../../settings/SettingsPane/SettingsPane';
-import DiagnosticsPane from '../../../diagnostic/DiagnosticPane';
+import DiagnosticsPane from '../../../diagnostic/DiagnosticPane/DiagnosticPane';
 import MemberSelector from '../MemberSelector/MemberSelector';
 import AppFooter from '../AppFooter/AppFooter';
