+++ src/components/ui/scaffold/MemberSelector/MemberSelector.js
@@ -5,9 +5,9 @@ import uuid from 'uuid';
 import PerformancePane from '../../../performance/PerformancePane/PerformancePane';
 import Neo4jConfiguration from '../../../configuration/Neo4jConfiguration/Neo4jConfiguration';
 import OSPane from '../../../performance/OSPane/OSPane';
-import PluginPane from '../../../db/PluginPane';
-import SampleQueryPane from '../../../db/SampleQueryPane';
-import LogsPane from '../../../db/LogsPane';
+import PluginPane from '../../../db/PluginPane/PluginPane';
+import SampleQueryPane from '../../../db/SampleQueryPane/SampleQueryPane';
+import LogsPane from '../../../db/LogsPane/LogsPane';
 import Tasks from '../../../performance/task/Tasks';
 import ClusterMemberMenuItem from '../ClusterMemberMenuItem/ClusterMemberMenuItem';
 import DetectedSettings from '../../../settings/DetectedSettings/DetectedSettings';
